Add files via upload
The purpose of this project is to look into different variables to observe their relationship, and plot a multiple linear regression based on several variables of individuals such as age, physical/family condition and location against their existing medical expense to be used for predicting future medical expenses of individuals that help medical insurance to make decision on charging the premium. Therefore, the expense is used as the response variable.